neon/requirements/requirements.txt
django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.2.4
markdown==2.6.5
django-haystack==2.4.1
django-phonenumber-field==1.0.0
python-dateutil==2.4.2
schedule==0.3.2
Pillow==3.1.0
twilio==5.3.0
django-localflavor==1.2
django-colorful==1.1.0
MySQL-python==1.2.5
django-filter==0.12.0
django-suit==0.2.17
django-rest-auth==0.7.0
requests==2.9.1
django-meta==0.3.2
simplejson==3.8.2
